travel_planner/requirements/base.txt
Django==5.0.3
psycopg2-binary==2.9.9
Pillow==10.2.0
django-environ==0.11.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.3
gunicorn==21.2.0
drf-yasg==1.21.7